IdentifyPoints {RgoogleMaps}	R Documentation

identify points by clicking on map

Description

The user can try to identify lat/lon pairs on the map by clicking on them

Usage

IdentifyPoints(MyMap, n = 1, verbose = 0)

Arguments

MyMap
map object

n
the maximum number of points to locate.

verbose
level of verbosity

Value

the lat/lon coordinates of the chosen points are returned

Author(s)

Markus Loecher

Examples

#The first step naturally will be to download a static map from the Google server. A simple example:

 #identifiy points:
 #IdentifyPoints(MyMap,5)

[Package RgoogleMaps version 1.2.0.7 Index]