MapBackground {RgoogleMaps}	R Documentation

get static Map from the Google server

Description

get static Map from the Google server

Usage

MapBackground(lat, lon, destfile, NEWMAP = TRUE, myTile, zoom = NULL, 
    size = c(640, 640), GRAYSCALE = FALSE, mar = c(0, 0, 0, 0), 
    PLOT = FALSE, verbose = 1, ...)

Arguments

lat

lon

destfile
File to load the map image from or save to, depending on NEWMAP.

NEWMAP
if TRUE, query the Google server and save to destfile, if FALSE load from destfile.

myTile
map tile from previous downloads

zoom
Google maps zoom level.

size
desired size of the map tile image. defaults to maximum size returned by the Gogle server, which is 640x640 pixels

GRAYSCALE
Boolean toggle; if TRUE the colored map tile is rendered into a black & white image, see RGB2GRAY

mar
outer margin in plot; if you want to see axes, change the default

PLOT
if TRUE, leave the plotting to PlotOnStaticMap, highly recommended

verbose
level of verbosity

...
further arguments to be passed to GetMap.bbox

Value

list containing the map tile

Author(s)

Markus Loecher

[Package RgoogleMaps version 1.2.0.7 Index]