MaxZoom {RgoogleMaps}	R Documentation

computes the maximum zoom level which will contain the given lat/lon range

Description

computes the maximum zoom level which will contain the given lat/lon range

Usage

MaxZoom(latrange, lonrange, size = c(640, 640))

Arguments

latrange
range of latitude values

lonrange
range of longitude values

size
desired size of the map tile image. defaults to maximum size returned by the Gogle server, which is 640x640 pixels

Value

zoom level

Author(s)

Markus Loecher

[Package RgoogleMaps version 1.2.0.7 Index]