NYleukemia {RgoogleMaps}	R Documentation

Upstate New York Leukemia Data

Description

Census tract level (n=281) leukemia data for the 8 counties in upstate New York from 1978-1982, paired with population data from the 1980 census.

Usage

data(NYleukemia)

Format

List with 5 items:

geo 	 table of the FIPS code, longitude, and latitude of the geographic centroid of each census tract

data 	 table of the FIPS code, number of cases, and population of each census tract

spatial.polygon 	 object of class SpatialPolygons (See SpatialPolygons-class) containing a map of the study region

surrounded 	 row IDs of the 4 census tracts that are completely surrounded by the surrounding census tracts

surrounding 	 row IDs of the 4 census tracts that completely surround the surrounded census tracts

Source

http://www.sph.emory.edu/~lwaller/ch4index.htm

References

Turnbull, B. W. et al (1990) Monitoring for clusters of disease: application to leukemia incidence in upstate New York American Journal of Epidemiology, 132, 136–143

Examples

data(NYleukemia)
  population <- NYleukemia$data$population
  cases <- NYleukemia$data$cases
  mapNY <- GetMap(center=c(lon=-76.00365, lat=42.67456), destfile = "NYstate.png",
  maptype = "mobile", zoom=9)
  ColorMap(100*cases/population, mapNY, NYleukemia$spatial.polygon, add = FALSE,
  alpha = 0.35, log = TRUE, location = "topleft")
 

[Package RgoogleMaps version 1.2.0.7 Index]