PlotArrowsOnStaticMap {RgoogleMaps}	R Documentation

plots arrows or segments on map

Description

This function plots/overlays arrows or segments on a map.

Usage

PlotArrowsOnStaticMap(MyMap, lat0, lon0, lat1 = lat0, lon1 = lon0, 
    TrueProj = TRUE, FUN = arrows, add = FALSE, verbose = 0, 
    ...)

Arguments

MyMap
map image returned from e.g. GetMap()

lat0
latitude valuesof points FROM which to draw.

lon0
longitude values of points FROM which to draw.

lat1
latitude valuesof points TO which to draw.

lon1
longitude values of points TO which to draw.

TrueProj
set to FALSE if you are willing to accept some degree of inaccuracy in the mapping. In that case, the coordinates of the image are in lat/lon and the user can simply overly points/lines/axis without worrying about projections

FUN
, plotting function to use for overlay; typical choices would be arrows and segments

add
start a new plot or add to an existing

verbose
level of verbosity

...
further arguments to be passed to FUN

Value

return value of FUN

Author(s)

Markus Loecher

See Also

PlotOnStaticMap arrows

Examples

    MyMap <- GetMap(center=c(lat=40.7,lon=-74), zoom=11)
PlotArrowsOnStaticMap(MyMap, lat0=40.69, lon0=-73.9, lat1=40.71, lon1=-74.1, col = 'red')   

[Package RgoogleMaps version 1.2.0.7 Index]