RGB2GRAY {RgoogleMaps}	R Documentation

translates an RGB image matrix to gray scale

Description

This function translates the rgb values of the array myTile into a scalar matrix with just one gray value per pixel.

Usage

RGB2GRAY(myTile)

Arguments

myTile
rgb image matrix, usually array with 3 dimensions

Details

Gray scale intensity defined as 0.30R + 0.59G + 0.11B

Value

image tile

Author(s)

Markus Loecher

Examples

if (interactive()){
  BrooklynLatLon = getGeoCode("Brooklyn")
  mapBrooklyn <- GetMap(center=BrooklynLatLon, destfile = file.path(tempdir(), "Brooklyn.png"), 
                 zoom=11, size = c(240,240))
  mapBrooklynBW$myTile = RGB2GRAY(mapBrooklyn$myTile)
  PlotOnStaticMap(mapBrooklynBW)
}

[Package RgoogleMaps version 1.2.0.7 Index]