ReadMapTile {RgoogleMaps}	R Documentation

Read a bitmap image stored in the PNG format

Description

Reads an image from a PNG file/content into a raster array.

Usage

ReadMapTile(destfile, METADATA = TRUE, native = TRUE)

Arguments

destfile
png file to read

METADATA
read MetaInfo as well ?

native
determines the image representation - if FALSE then the result is an array, if TRUE then the result is a native raster representation, see readPNG in package png.

Value

map or tile object

Author(s)

Markus Loecher

[Package RgoogleMaps version 1.2.0.7 Index]