RgoogleMaps-package {RgoogleMaps}	R Documentation

Overlays on Google map tiles in R

Description

This package serves two purposes: (i) Provide a comfortable R interface to query the Google server for static maps, and (ii) Use the map as a background image to overlay plots within R. This requires proper coordinate scaling.

Details

Package: 	 RgoogleMaps

Type: 	 Package

Title: 	 Overlays on Google map tiles in R

Version: 	 1.2.0.7

Date: 	 2015-01-20

Depends: 	 R (>= 2.10)

Imports: 	 graphics, stats, utils, png, RJSONIO

Suggests: 	 PBSmapping, maptools, sp, rgdal, loa, RColorBrewer

Author: 	 Markus Loecher

Maintainer: 	 Markus Loecher <markus.loecher@gmail.com>

License: 	 GPL

LazyLoad: 	 yes

Author(s)

Markus Loecher

[Package RgoogleMaps version 1.2.0.7 Index]